Add -PassThru support for methods to persist to disk (#32)
* Add -PassThru to Get-BloggerPost #26

* Added -PassThru to ConvertTo-HtmlFromMarkdown and tests

* Add -PassThru to ConvertTo-MarkdownFromHtml

+++ src/public/ConvertTo-HtmlFromMarkdown.ps1
@@ -8,20 +8,38 @@
 .PARAMETER OutFile
   The resulting html. If this parameter is not specified an HTML file with the same name of the markdown file will be created.
 
+.PARAMETER PassThru
+  If specified, the content of the HTML file will be returned as well as written to disk.
+
+.EXAMPLE
+  # obtain an HTML representation of the markdown file
+  $html = ConvertTo-HtmlFromMarkdown -File "C:\path\to\file.md"
+
+.EXAMPLE
+  # write the HTML representation of the markdown file to disk
+  ConvertTo-HtmlFromMarkdown -File "C:\path\to\file.md" -OutFile "C:\path\to\file.html"
+
+.EXAMPLE
+  # write the HTML representation of the markdown file to disk and return the content
+  $html = ConvertTo-HtmlFromMarkdown -File "C:\path\to\file.md" -OutFile "C:\path\to\file.html" -PassThru
 #>
 function ConvertTo-HtmlFromMarkdown {
+  [CmdletBinding(DefaultParameterSetName = "Default")]
   param(
-    [Parameter(Mandatory = $true, HelpMessage = "Path to Markdown file")]
+    [Parameter(Mandatory = $true, ParameterSetName = "Default")]
+    [Parameter(Mandatory = $true, ParameterSetName = "Persist")]
     [ValidateScript({ Test-Path $_ -PathType Leaf })]
     [string]$File,
 
-    [Parameter(HelpMessage = "File path to create")]
-    #[ValidateScript({ Test-Path $_ -Include "*.html" -PathType Container})]
-    [string]$OutFile
+    [Parameter(ParameterSetName = "Persist")]
+    [string]$OutFile,
+
+    [Parameter(ParameterSetName = "Persist")]
+    [switch]$PassThru
   )
 
   # ensure that the file is an absolute path because pandoc.exe doesn't like powershell relative paths
-  $File = (Resolve-Path $File).Path
+  $File = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($File)
 
   # Use pandoc to convert the markdown to Html 
   $pandocArgs = "`"{0}`" " -f $File
@@ -41,8 +59,10 @@ function ConvertTo-HtmlFromMarkdown {
 
   if (!($OutFile)) {
     $OutFile = Join-Path (Split-Path $File -Parent) ((Split-Path $File -LeafBase) + ".html")
+    # ensure that the file is an absolute path because pandoc.exe doesn't like powershell relative paths
     Write-Verbose "Using OutFile: $OutFile"
   }
+  $OutFile = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($OutFile)
 
   $pandocArgs += "-o `"{0}`" " -f $OutFile
 
@@ -60,7 +80,14 @@ function ConvertTo-HtmlFromMarkdown {
 
   Set-Content -Path $OutFile -Value $content
 
-  Remove-Item $OutFile
+  if (!($PSCmdlet.ParameterSetName -eq "Persist")) {
+    Write-Verbose "Removing temporary file: $OutFile"
+    Remove-Item $OutFile
 
-  return $content
+    return $content
+  }
+
+  if ($PassThru.IsPresent -and $PassThru) {
+    return $content
+  }
 }
